refactor to split 2d operations out into separate library file
